requirements.txt
fastapi[all]
uvicorn[standard]
accelerate
torch
torchvision
pillow
opencv-python
transformers
diffusers